+++ test/td_audit/subscriptions/subscriptions_test.exs
@@ -2,10 +2,10 @@ defmodule TdAudit.SubscriptionsTest do
   @moduledoc """
   Subscriptions testing module
   """
-  alias TdAudit.Audit
   use TdAudit.DataCase
+
+  alias TdAudit.Audit
   alias TdAudit.Subscriptions
-  import TdAudit.SubscriptionTestHelper
 
   describe "subscriptions" do
     alias TdAudit.Subscriptions.Subscription
@@ -31,46 +31,46 @@ defmodule TdAudit.SubscriptionsTest do
     end
 
     test "list_subscriptions/0 returns all subscriptions" do
-      subscription = subscription_fixture()
+      subscription = insert(:subscription)
       assert Subscriptions.list_subscriptions() == [subscription]
     end
 
     test "list_subscriptions/1 returns all subscriptions filtered by resource_id" do
       target_resource_id = 44
-      subscription_fixture()
-      subscription_1 = subscription_fixture(%{resource_id: target_resource_id, event: "my event"})
-      subscription_2 = subscription_fixture(%{resource_id: target_resource_id, event: "my event"})
+      insert(:subscription)
+      s1 = insert(:subscription, resource_id: 44, event: "my event")
+      s2 = insert(:subscription, resource_id: 44, event: "my event")
 
       assert Subscriptions.list_subscriptions_by_filter(%{
                "resource_id" => target_resource_id,
                "event" => ["my event"]
-             }) == [subscription_1, subscription_2]
+             }) == [s1, s2]
     end
 
     test "get_subscription!/1 returns the subscription with given id" do
-      subscription = subscription_fixture()
+      subscription = insert(:subscription)
       assert Subscriptions.get_subscription!(subscription.id) == subscription
     end
 
     test "create_subscription/1 with valid data creates a event" do
-      assert {:ok, %Subscription{} = subscription} =
-               Subscriptions.create_subscription(retrieve_valid_attrs())
+      fields = [:event, :resource_id, :resource_type, :user_email, :periodicity]
+      params = build(:subscription) |> Map.take(fields)
 
-      assert subscription.event == "some event"
-      assert subscription.resource_id == 42
-      assert subscription.resource_type == "some resource_type"
-      assert subscription.user_email == "mymail@foo.com"
+      assert {:ok, %Subscription{} = subscription} = Subscriptions.create_subscription(params)
+
+      assert Map.take(subscription, fields) == params
     end
 
     test "create_subscription/1 with invalid data returns error changeset" do
-      assert {:error, %Ecto.Changeset{}} =
-               Subscriptions.create_subscription(retrieve_invalid_attrs())
+      params = build(:subscription) |> Map.from_struct() |> Map.delete(:user_email)
+
+      assert {:error, %Ecto.Changeset{}} = Subscriptions.create_subscription(params)
     end
 
     test "update_subscription/2 with valid data updates the subscription" do
       new_resource_id = 44
       new_resource_type = "new resource type"
-      subscription = subscription_fixture()
+      subscription = insert(:subscription)
 
       assert {:ok, subscription} =
                Subscriptions.update_subscription(subscription, %{
@@ -84,7 +84,7 @@ defmodule TdAudit.SubscriptionsTest do
     end
 
     test "update_subscription/2 with invalid data returns error changeset" do
-      subscription = subscription_fixture()
+      subscription = insert(:subscription)
 
       assert {:error, %Ecto.Changeset{}} =
                Subscriptions.update_subscription(subscription, %{"resource_id" => nil})
@@ -93,14 +93,9 @@ defmodule TdAudit.SubscriptionsTest do
     end
 
     test "delete_subscription/1 deletes the subscription" do
-      subscription = subscription_fixture()
+      subscription = insert(:subscription)
       assert {:ok, %Subscription{}} = Subscriptions.delete_subscription(subscription)
       assert_raise Ecto.NoResultsError, fn -> Subscriptions.get_subscription!(subscription.id) end
-    end
-
-    test "change_subscription/1 returns a subscription changeset" do
-      subscription = subscription_fixture()
-      assert %Ecto.Changeset{} = Subscriptions.change_subscription(subscription)
     end
   end
 end